Move dummy viewer tests into separate file
