+++ tests/test_astro_image_display_api.py
@@ -1,8 +1,6 @@
 from astropy.utils.data import get_pkg_data_contents
 
 from astro_image_display_api.interface_definition import ImageViewerInterface
-from astro_image_display_api.widget_api_test import ImageWidgetAPITest
-from astro_image_display_api.dummy_viewer import ImageViewer
 
 def test_api_test_class_completeness():
     """
@@ -28,12 +26,3 @@ def test_api_test_class_completeness():
         "ImageWidgetAPITest does not access these attributes/methods:\n "
         f"{"\n".join(attr for attr, present in zip(required_attributes, attr_present) if not present)}. "
     )
-
-
-def test_instance():
-    image = ImageViewer()
-    assert isinstance(image, ImageViewerInterface)
-
-
-class TestDummyViewer(ImageWidgetAPITest):
-    image_widget_class = ImageViewer
